[TEST] Split the tests into internal and public
Move all the tests that reference the internal members of CsvConverter to a separate project. Make the internal members visible to it only. This allows a separate test project to use the same test class files as the public test project and the NuGet package rather than a project dependency.
